requirements.txt
# Core Web Crawling (consolidated from requirements_core.txt and requirements_minimal.txt)
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.2

# PDF Processing (tika from core, PyPDF2 for bond crawler)
tika>=1.24
PyPDF2>=3.0.0

# Korean NLP
ekonlpy>=0.0.97
pykospacing>=0.5

# Data Processing
pandas>=2.0.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
wordcloud>=1.8.0

# Topic Modeling
gensim==4.3.2

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dateutil==2.8.2
jupyter==1.0.0
notebook==7.0.2

# Text Processing
regex==2023.8.8

# Additional dependencies for Scrapy (updated from core requirements)
itemadapter>=0.8.0
twisted>=20.3.0

# For concurrent processing
# concurrent-futures is built-in since Python 3.2+

# Environment variables management
python-dotenv>=1.0.0 